refactor: reorganize namespaces and file structure for Wangkanai.Audit
