+++ Audit/src/Application/Configurations/AuditableConfigurationBuilder.cs
@@ -2,7 +2,7 @@
 
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-namespace Wangkanai.Audit.Domain.Configurations;
+namespace Wangkanai.Audit.Configurations;
 
 public static class AuditableConfigurationBuilder
 {
