+++ Audit/src/Application/Wangkanai.Audit.Application.csproj
@@ -1,13 +1,11 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFramework>net9.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
+   <ItemGroup>
+      <PackageReference Include="Microsoft.Extensions.Identity.Stores"/>
+   </ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\Domain\Wangkanai.Audit.Domain.csproj" />
-  </ItemGroup>
+   <ItemGroup>
+      <ProjectReference Include="..\Domain\Wangkanai.Audit.Domain.csproj"/>
+   </ItemGroup>
 
 </Project>
