+++ Audit/src/Domain/Extensions/TrailExtensions.cs
@@ -1,5 +1,5 @@
 // Copyright (c) 2014-2025 Sarin Na Wangkanai, All Rights Reserved.
 
-namespace Wangkanai.Audit.Domain;
+namespace Wangkanai.Audit;
 
 public static class TrailExtensions { }
