+++ Audit/src/Domain/Interfaces/ICreatedEntity.cs
@@ -1,6 +1,6 @@
 // Copyright (c) 2014-2025 Sarin Na Wangkanai, All Rights Reserved.
 
-namespace Wangkanai.Audit.Domain;
+namespace Wangkanai.Audit;
 
 /// <summary>Represents an entity that tracks the creation date and time with timezone awareness.</summary>
 /// <remarks>
