+++ Audit/src/Domain/Interfaces/IDeletable.cs
@@ -1,6 +1,6 @@
 // Copyright (c) 2014-2025 Sarin Na Wangkanai, All Rights Reserved.
 
-namespace Wangkanai.Audit.Domain;
+namespace Wangkanai.Audit;
 
 /// <summary>Represents an entity that supports tracking its deletable state.</summary>
 /// <remarks>
