+++ Audit/src/Domain/Interfaces/IUserDeletedAuditableEntity.cs
@@ -1,6 +1,6 @@
 // Copyright (c) 2014-2025 Sarin Na Wangkanai, All Rights Reserved.
 
-namespace Wangkanai.Audit.Domain;
+namespace Wangkanai.Audit;
 
 /// <summary>
 /// Defines an interface for entities that track deletion actions,
