+++ Audit/src/Domain/Stores/IQueryableTrailStore.cs
@@ -1,6 +1,6 @@
 // Copyright (c) 2014-2025 Sarin Na Wangkanai, All Rights Reserved.
 
-namespace Wangkanai.Audit.Domain;
+namespace Wangkanai.Audit;
 
 /// <summary>
 /// Represents an abstraction for a queryable audit trail store, providing access to audit trail records for tracking changes in the system.
