+++ Audit/src/Domain/Trail.cs
@@ -6,7 +6,7 @@ using System.Text.Json.Serialization;
 
 using Wangkanai.Foundation;
 
-namespace Wangkanai.Audit.Domain;
+namespace Wangkanai.Audit;
 
 /// <summary>Represents an audit trail record for tracking entity changes in the system.</summary>
 /// <typeparam name="TKey">The type of the unique identifier for the audit trail.</typeparam>
