+++ Audit/src/Domain/TrailType.cs
@@ -1,6 +1,6 @@
 // Copyright (c) 2014-2025 Sarin Na Wangkanai, All Rights Reserved.
 
-namespace Wangkanai.Audit.Domain;
+namespace Wangkanai.Audit;
 
 /// <summary>Specifies the type of operation recorded in an audit trail.</summary>
 public enum TrailType : byte
