+++ Audit/src/Domain/UserAuditableEntity.cs
@@ -1,6 +1,6 @@
 ﻿// Copyright (c) 2014-2025 Sarin Na Wangkanai, All Rights Reserved.
 
-namespace Wangkanai.Audit.Domain;
+namespace Wangkanai.Audit;
 
 /// <summary>
 /// Represents an abstract base class for auditable entities that captures user-related information during create and update operations.
