+++ Audit/src/Domain/Usings.cs
@@ -1,5 +1,3 @@
 // Copyright (c) 2014-2025 Sarin Na Wangkanai, All Rights Reserved.
 
 global using Microsoft.AspNetCore.Identity;
-
-global using Wangkanai.Domain;
