+++ Foundation/src/Application/Configurations/EntityConfiguration.cs
@@ -1,5 +1,6 @@
 // Copyright (c) 2014-2025 Sarin Na Wangkanai, All Rights Reserved.
 
+using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
 namespace Wangkanai.Foundation.Configurations;
