+++ Foundation/src/Application/Configurations/KeyConfigurationBuilder.cs
