+++ Foundation/src/Application/Configurations/VersionConfigurationBuilder.cs
