+++ Foundation/src/Application/Wangkanai.Foundation.Application.csproj
@@ -1,13 +1,15 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-   <PropertyGroup>
-      <TargetFramework>net9.0</TargetFramework>
-      <ImplicitUsings>enable</ImplicitUsings>
-      <Nullable>enable</Nullable>
-   </PropertyGroup>
+   <ItemGroup>
+      <PackageReference Include="Microsoft.Extensions.Identity.Stores"/>
+   </ItemGroup>
 
    <ItemGroup>
-      <ProjectReference Include="..\Domain\Wangkanai.Foundation.Domain.csproj" />
+      <ProjectReference Include="..\Domain\Wangkanai.Foundation.Domain.csproj"/>
+   </ItemGroup>
+
+   <ItemGroup>
+      <Folder Include="Extensions\"/>
    </ItemGroup>
 
 </Project>
