+++ Foundation/src/Domain/Usings.cs
@@ -2,11 +2,5 @@
 
 global using System;
 global using System.Collections.Generic;
-global using System.ComponentModel.DataAnnotations;
 global using System.Diagnostics.CodeAnalysis;
 global using System.Linq;
-
-global using Microsoft.AspNetCore.Identity;
-global using Microsoft.EntityFrameworkCore;
-
-global using System.Diagnostics;
