+++ Foundation/src/Domain/Wangkanai.Foundation.Domain.csproj
@@ -4,7 +4,7 @@
 
    </PropertyGroup>
    <ItemGroup>
-      <PackageReference Include="Microsoft.Extensions.Identity.Stores" />
+
       <PackageReference Include="Microsoft.EntityFrameworkCore" />
       <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" />
       <PackageReference Include="Wangkanai.System" />
