+++ Foundation/tests/Application/Configurations/VersionConfigurationBuilderTests.cs
