+++ Foundation/tests/Domain/Configurations/Cases/IntEntityTypeConfiguration.cs
@@ -3,6 +3,8 @@
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
+using Wangkanai.Foundation.Models;
+
 namespace Wangkanai.Foundation.Configurations.Cases;
 
 public class IntEntityTypeConfiguration : IEntityTypeConfiguration<IntEntity>
