+++ Foundation/tests/Domain/EntityTests.cs
@@ -1,5 +1,7 @@
 ﻿// Copyright (c) 2014-2025 Sarin Na Wangkanai, All Rights Reserved.
 
+using Wangkanai.Foundation.Models;
+
 namespace Wangkanai.Foundation;
 
 public class EntityTests
